speed up Element.ByRuntimeId searches by utilizing a closest parent
Before, Element.ByRuntimeId was done by searching through the tree by only using the RuntimeId as a criteria. Now, the UiaDll will try to take a RuntimeId like [42 12345 0 1] and translate each part as a Window handle if it can. If so, it will scope the search to its closest parent (i.e. window handle 12345 in this case).